requirements.txt
request=2.31.0
metaphor_python=0.1.17